% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longestname.R
\name{longestname}
\alias{longestname}
\title{Longest First Name}
\usage{
longestname( year )
}
\arguments{
\item{year}{Sets the year of graduates to be scanned}
}
\value{
the name of graduating senior with the longest first name
}
\description{
Shows the name of graduating senior of \code{year} with the longest first name
}

