% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readgraduates.R
\name{readgraduates}
\alias{readgraduates}
\title{Read Graduates}
\usage{
readgraduates( year )
}
\arguments{
\item{year}{Sets the year of graduates to be scanned}
}
\value{
A list of names of graduates from \code{year}
}
\description{
Scan for the names of graduates and create a list of names
}

