% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showstats.R
\name{showstats}
\alias{showstats}
\title{Show Statistics}
\usage{
showstats(type)
}
\arguments{
\item{type}{The kind of analysis to be done. The possible options are \code{graduates}, \code{SCL}, \code{MCL}, \code{CL} and \code{summary}. \code{graduates}, \code{SCL}, \code{MCL} and \code{CL} displays graph of the total number of graduating seniors, those with Cum Laude or higher, those with Magna Cum Laude or hight and those with Summa Cum Laude respectively over the years. \code{summary} shows a graph summarizing the graduating seniors over the years.}
}
\value{
Results from the kind of analysis specified by \code{type}.
}
\description{
Uses the data from the read function and performs appropriate analysis and generates graphics associated with the data.
}

